+++ src/shared/hooks/useDebounce.ts
@@ -1,6 +1,6 @@
-import { DebounceType } from '@dev/shared/types/common/debounce.type';
+import { DebounceType } from '@dev/shared/types/debounce.type';
 import { useCallback, useState } from 'react';
-import { debounce } from '@devUtils/common/debounce';
+import { debounce } from '@devUtils/debounce';
 
 export const useDebounce = ({ fn, wait = 500 }: DebounceType) => {
     // 상태 관리 (선택 사항, 필요하다면 사용할 수 있음)
