+++ src/shared/types/throttle.type.ts
