+++ src/shared/utils/error.ts
