+++ tailwind.config.ts
@@ -2,8 +2,8 @@ import type {Config} from "tailwindcss";
 
 const config: Config = {
     content: [
-        "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
-        "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
+        "./src/views/**/*.{js,ts,jsx,tsx,mdx}",
+        "./src/widgets/**/*.{js,ts,jsx,tsx,mdx}",
         "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
     ],
     theme: {
